airbnb_planner_multiagent/weather_agent/requirements.txt
a2a-sdk>=0.2.9
httpx>=0.28.1
httpx-sse>=0.4.0
jwcrypto>=1.5.6
pydantic>=2.10.6
pyjwt>=2.10.1
sse-starlette>=2.2.1
starlette>=0.46.1
python-dotenv>=1.0.1
typing-extensions>=4.12.2
uvicorn>=0.34.0
click>=8.2.0
geopy>=2.4.1
google-adk>=1.0.0
mcp>=1.5.0
langchain-core>=0.2.29